타입 분리2(#164)
