+++ apps/web/features/mypage/edit/types/index.ts
@@ -1,0 +1,18 @@
+export interface GetEditProfileParams {
+  userId: string;
+}
+
+export interface UseGetEditProfileParams {
+  userId: string;
+}
+
+export interface ProfilePreviewProps {
+  profileImg?: string | null;
+  onChangeAvatar?: (avatar: UploadedImage) => void;
+  onDelete?: () => void;
+}
+
+export interface UploadedImage {
+  file: File;
+  preview: string;
+}
