+++ apps/web/features/mypage/ui/MyPageMenuList.tsx
@@ -2,10 +2,7 @@ import StatusBadge from '@/shared/ui/badge/StatusBadge';
 import { Heart, MapPin } from 'lucide-react';
 import { Button } from '@repo/ui/components/Button/Button';
 import MyPageMenuItem from '@/features/mypage/ui/MyPageMenuItem';
-
-interface MyPageMenuListProps {
-  address: string;
-}
+import { MyPageMenuListProps } from '@/features/mypage/types';
 
 const MyPageMenuList = ({ address }: MyPageMenuListProps) => {
   return (
